# [DataFrame] Read files in parallel (4x faster)