# Which issue does this PR close?

Closes https://github.com/apache/arrow-datafusion/issues/6983 
Closes https://github.com/apache/arrow-datafusion/discussions/6908

# Rationale for this change

This code uses a single core to read the file
```
    let _df = _ctx.read_parquet(FILENAME, _read_options).await.unwrap();
    let _cached = _df.cache().await;
```

# What changes are included in this PR?

Use multiple cores

Testing using using `cargo --release`

With main (16s)
```
datafusion end -> 2023-07-16T09:07:29.895269-04:00 16.080133858s
```

With this branch (3s)
```
datafusion end -> 2023-07-16T08:52:05.511984-04:00 2.947019517s
```

# Are these changes tested?
Yes
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

# Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->