@alamb 2023-07-17
I think a better approach is described on https://github.com/apache/arrow-datafusion/issues/6983#issuecomment-1638250822